# server/swagger/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gateway MERN (MongoDB, ExpressJS, React, NodeJS).",
    "description": "Simple and extensible API for managing gateways.",
    "contact": {
      "email": "adriangonzalezsanchez1996@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/device/removeFromGateway/{id}": {
      "get": {
        "tags": [
          "device"
        ],
        "summary": "Remove this device from it's gateway device list",
        "operationId": "removeDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the device",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device was successfully removed"
          },
          "400": {
            "description": "Device was not Found"
          }
        }
      }
    },
    "/device/addToGateway/{id}": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "Add this device to the suplied gateway's devices list",
        "operationId": "addDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the device",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Gateway id",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device was successfully added to the gateway's list."
          },
          "400": {
            "description": "Device was not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Payload": {
          "type": "object",
          "properties": {
              "id": {
                  "type": "integer"
              }
          }
      },
...
    }
  }
}